Make the Transactor wait for tx to be mined (#428)
As a bit more context this again to make the transaction be more modular and allow other use cases and different implementation
