+++ pkg/common/contractInterface.go
@@ -7,12 +7,13 @@ import (
 	"context"
 	"math/big"
 
+	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/core/types"
 )
 
 // TransactionGeneratorFN is a callback function that a transaction submitter uses to actually invoke
 // a contract and generate a transaction.
-type TransactionGeneratorFN func(ctx context.Context, contract ContractInterface) (*types.Transaction, error)
+type TransactionGeneratorFN func(ctx context.Context, options *bind.TransactOpts) (*types.Transaction, error)
 
 // ContractInterface represents an abstraction of function definitions that can be
 // called on the smart contract. This is mostly so that we can do unit tests without
